requirements.txt
PyQt6==6.10.1
PyQt6-Qt==6.0.1
PyQt6-Qt6==6.10.1
PyQt6.Webengine==6.10.0
PyQt6-WebEngine-Qt6==6.10.1
PyQt6-Sip
websockets
psutil
pypdfium2
pillow
requests
urllib3
